6.10.x validate hostname

+++ tasks/go.py
@@ -2,14 +2,16 @@
 Golang related tasks go here
 """
 from __future__ import print_function
+import datetime
 import os
+import shutil
 import sys
-import json
 
 from invoke import task
 from invoke.exceptions import Exit
 from .build_tags import get_default_build_tags
 from .utils import pkg_config_path, get_build_flags
+from .bootstrap import get_deps, process_deps
 
 
 # List of modules to ignore when running lint on Windows platform
@@ -19,6 +21,7 @@ WIN_MODULE_WHITELIST = [
     "iostats_wmi_windows.go",
     "pdh.go",
     "pdhhelper.go",
+    "shutil.go",
     "tailer_windows.go",
 ]
 
@@ -26,19 +29,8 @@ WIN_MODULE_WHITELIST = [
 MISSPELL_IGNORED_TARGETS = [
     os.path.join("cmd", "agent", "dist", "checks", "prometheus_check"),
     os.path.join("cmd", "agent", "gui", "views", "private"),
+    os.path.join("pkg", "collector", "corechecks", "system", "testfiles"),
 ]
-
-# Bootstrap dependencies description
-BOOTSTRAP_DEPS = "bootstrap.json"
-
-
-def get_deps(key):
-    here = os.path.abspath(os.path.dirname(__file__))
-    with open(os.path.join(here, '..', BOOTSTRAP_DEPS)) as f:
-        deps = json.load(f)
-
-    return deps.get(key, {})
-
 
 @task
 def fmt(ctx, targets, fail_on_fmt=False):
@@ -117,6 +109,7 @@ def vet(ctx, targets, use_embedded_libs=False):
     # add the /... suffix to the targets
     args = ["{}/...".format(t) for t in targets]
     build_tags = get_default_build_tags()
+    build_tags.append("novet")
 
     _, _, env = get_build_flags(ctx, use_embedded_libs=use_embedded_libs)
 
@@ -192,48 +185,69 @@ def misspell(ctx, targets):
         print("misspell found no issues")
 
 @task
-def deps(ctx, no_checks=False, core_dir=None, verbose=False):
+def deps(ctx, no_checks=False, core_dir=None, verbose=False, android=False):
     """
     Setup Go dependencies
     """
     verbosity = ' -v' if verbose else ''
     deps = get_deps('deps')
-    for tool, version in deps.iteritems():
-        # download tools
-        path = os.path.join(os.environ.get('GOPATH'), 'src', tool)
-        if not os.path.exists(path):
-            ctx.run("go get{} -d -u {}".format(verbosity, tool))
+    order = deps.get("order", deps.keys())
+    for dependency in order:
+        tool = deps.get(dependency)
+        if not tool:
+            print("Malformed bootstrap JSON, dependency {} not found". format(dependency))
+            raise Exit(code=1)
+        print("processing checkout tool {}".format(dependency))
+        process_deps(ctx, dependency, tool.get('version'), tool.get('type'), 'checkout', verbose=verbose)
 
-        with ctx.cd(path):
-            # checkout versions
-            ctx.run("git fetch")
-            ctx.run("git checkout {}".format(version))
+    order = deps.get("order", deps.keys())
+    for dependency in order:
+        tool = deps.get(dependency)
+        print("processing get tool {}".format(dependency))
+        process_deps(ctx, dependency, tool.get('version'), tool.get('type'), 'install', verbose=verbose)
 
-    postdeps = get_deps('post-deps')
-    for tool, version in postdeps.iteritems():
-        # download tools
-        path = os.path.join(os.environ.get('GOPATH'), 'src', tool)
-        if not os.path.exists(path):
-            ctx.run("go get{} -d -u {}".format(verbosity, tool))
+    if android:
+        ndkhome = os.environ.get('ANDROID_NDK_HOME')
+        if not ndkhome:
+            print("set ANDROID_NDK_HOME to build android")
+            raise Exit(code=1)
 
-        with ctx.cd(path):
-            # checkout versions
-            ctx.run("git fetch")
-            ctx.run("git checkout {}".format(version))
-
-    for tool, version in deps.iteritems():
-        # install tools
-        ctx.run("go install{} {}".format(verbosity, tool))
+        cmd = "gomobile init -ndk {}". format(ndkhome)
+        print("gomobile command {}". format(cmd))
+        ctx.run(cmd)
 
     # source level deps
+    print("calling dep ensure")
+    start = datetime.datetime.now()
     ctx.run("dep ensure{}".format(verbosity))
+    dep_done = datetime.datetime.now()
+    
+    # If github.com/DataDog/datadog-agent gets vendored too - nuke it
+    #
+    # This may happen as a result of having to introduce DEPPROJECTROOT
+    # in our builders to get around a known-issue with go dep, and the
+    # strange GOPATH situation in our builders.
+    #
+    # This is only a workaround, we should eliminate the need to resort
+    # to DEPPROJECTROOT.
+    if os.path.exists('vendor/github.com/DataDog/datadog-agent'):
+        print("Removing vendored github.com/DataDog/datadog-agent")
+        shutil.rmtree('vendor/github.com/DataDog/datadog-agent')
+
     # make sure PSUTIL is gone on windows; the dep ensure above will vendor it
     # in because it's necessary on other platforms
-    if sys.platform == 'win32':
+    if not android and sys.platform == 'win32':
         print("Removing PSUTIL on Windows")
         ctx.run("rd /s/q vendor\\github.com\\shirou\\gopsutil")
 
+    # Make sure that golang.org/x/mobile is deleted.  It will get vendored in
+    # because we use it, and there's no way to exclude; however, we must use
+    # the version from $GOPATH
+    if os.path.exists('vendor/golang.org/x/mobile'):
+        print("Removing vendored golang.org/x/mobile")
+        shutil.rmtree('vendor/golang.org/x/mobile')
 
+    checks_start = datetime.datetime.now()
     if not no_checks:
         verbosity = 'v' if verbose else 'q'
         core_dir = core_dir or os.getenv('DD_CORE_DIR')
@@ -241,13 +255,59 @@ def deps(ctx, no_checks=False, core_dir=None, verbose=False):
         if core_dir:
             checks_base = os.path.join(os.path.abspath(core_dir), 'datadog_checks_base')
             ctx.run('pip install -{} -e {}'.format(verbosity, checks_base))
+            ctx.run('pip install -{} -r {}'.format(verbosity, os.path.join(checks_base, 'requirements.in')))
         else:
             core_dir = os.path.join(os.getcwd(), 'vendor', 'integrations-core')
             checks_base = os.path.join(core_dir, 'datadog_checks_base')
             if not os.path.isdir(core_dir):
                 ctx.run('git clone -{} https://github.com/DataDog/integrations-core {}'.format(verbosity, core_dir))
             ctx.run('pip install -{} {}'.format(verbosity, checks_base))
+            ctx.run('pip install -{} -r {}'.format(verbosity, os.path.join(checks_base, 'requirements.in')))
+    checks_done = datetime.datetime.now()
 
+    print("dep ensure, elapsed:    {}".format(dep_done - start))
+    print("checks install elapsed: {}".format(checks_done - checks_start))
+
+@task
+def lint_licenses(ctx):
+    """
+    Checks that the LICENSE-3rdparty.csv file is up-to-date with contents of Gopkg.lock
+    """
+    import csv
+    import toml
+
+    # non-go deps that should be listed in the license file, but not in Gopkg.lock
+    NON_GO_DEPS = set([
+        'github.com/codemirror/CodeMirror',
+        'github.com/FortAwesome/Font-Awesome',
+        'github.com/jquery/jquery',
+    ])
+
+    # Read all dep names from Gopkg.lock
+    go_deps = set()
+    gopkg_lock = toml.load('Gopkg.lock')
+    for project in gopkg_lock['projects']:
+        # FIXME: this conditional is necessary because of the issue introduced by DEPPROJECTROOT
+        # (for some reason `datadog-agent` gets added to Gopkg.lock and vendored), see comment in `deps`
+        # task for details
+        if project['name'] != 'github.com/DataDog/datadog-agent':
+            go_deps.add(project['name'])
+
+    deps = go_deps | NON_GO_DEPS
+
+    # Read all dep names listed in LICENSE-3rdparty
+    licenses = csv.DictReader(open('LICENSE-3rdparty.csv'))
+    license_deps = set()
+    for entry in licenses:
+        if len(entry['License']) == 0:
+            raise Exit(message="LICENSE-3rdparty entry '{}' has an empty license".format(entry['Origin']), code=1)
+        license_deps.add(entry['Origin'])
+
+    if deps != license_deps:
+        raise Exit(message="LICENSE-3rdparty.csv is outdated compared to deps listed in Gopkg.lock:\n" +
+                           "missing from LICENSE-3rdparty.csv: {}\n".format(deps - license_deps) +
+                           "listed in LICENSE-3rdparty.csv but not in Gopkg.lock: {}".format(license_deps - deps),
+                   code=1)
 
 @task
 def reset(ctx):
